Make make_nvp member of archives
There are instances where, in a given compilation unit, we never want
to serialize a serializable type but simply use it. If the type has a
serialize() member that uses ::cereal::make_nvp however, the cereal
header must always be included.

With this change we propose adding make_nvp to the archives themselves
so a serializable type is usable without including any cereal headers.
But at the same time it is possible to maintain the serialization code
as a member of the type, which is preferable over external definition
of the serialization code in a separate class.